queue: split struct QueueState into submodule
Split `struct QueueState` into a submodule.
